+++ backend/src/baserow/core/handler.py
@@ -146,6 +146,7 @@ class CoreHandler:
                 "allow_new_signups",
                 "allow_signups_via_group_invitations",
                 "allow_reset_password",
+                "allow_global_group_creation",
                 "account_deletion_grace_delay",
             ],
             settings_instance,
@@ -402,11 +403,9 @@ class CoreHandler:
         :return: The newly created GroupUser object
         """
 
-        group = Group.objects.create(name=name)
+        CoreHandler().check_permissions(user, CreateGroupOperationType.type)
 
-        CoreHandler().check_permissions(
-            user, CreateGroupOperationType.type, group=group
-        )
+        group = Group.objects.create(name=name)
 
         last_order = GroupUser.get_last_order(user)
         group_user = GroupUser.objects.create(